+++ tests/test_util.py
@@ -1,3 +1,5 @@
+# pylint: skip-file
+
 from repopulator.util import *
 
 def test_lowerBound():