+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+### Added
+- CRUD for all repo-related objects
+
+### Fixed
+- PgpSigner now respects home directory parameter
+- Project is now Stable
+
 ## [0.7] - 2024-06-05
 
 ### Changed